pacuba/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-paypal==1.0.0
django-widget-tweaks==1.4.8
idna==2.10
importlib-metadata==2.1.1
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
zipp==3.4.0
